+++ changelog.md
@@ -5,4 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 1.1.0 - 2025-11-17
+
+### Added
+
+- The cost can be rounded to a whole number, or to 2, 3, or 4 decimal places, in accordance
+  with [ISO 4217](https://www.iso.org/iso-4217-currency-codes.html).
+
+### Changed
+
+- Update dependencies and plugins: jUnit, JAR.
+
+## 1.0.0 - 2025-10-28
+
+### Initial release
+
+- The cost can be calculated from the price, the distance, and either the distance per unit of fuel or the fuel used per
+  unit of distance.